Fix x86_64 build-install Compliance
-> Fix hardcoded path
-> MmsPluginStorage.cpp: change variable type from unsinged int to size_t to avoid:
cannot convert 'unsigned int*' to 'size_t* {aka long unsigned int*}' for argument '2' to 'char* _MsgMmsSerializeMessageData(const MMS_MESSAGE_DATA_S*, size_t*)'

+++ plugin/mms_plugin/MmsPluginStorage.cpp
@@ -478,7 +478,7 @@ msg_error_t	MmsPluginStorage::plgGetMmsMessage(MSG_MESSAGE_INFO_S *pMsg, MSG_SEN
 	MMS_MESSAGE_DATA_S tempMmsMsg = {0,};
 	MMS_MESSAGE_DATA_S *pMmsMsg = &tempMmsMsg;
 	int partCnt = 0;
-	unsigned int nSize = 0;
+	size_t nSize = 0;
 
 	MsgType partHeader;
 	MmsAttrib pMmsAttrib;
@@ -1425,8 +1425,8 @@ msg_error_t MmsPluginStorage::plgGetMmsMessage(MSG_MESSAGE_INFO_S *pMsg, MSG_SEN
 	msg_error_t	err = MSG_SUCCESS;
 	MMS_MESSAGE_DATA_S tempMmsMsgData = {0,};
 	MMS_MESSAGE_DATA_S *pMmsMsg = &tempMmsMsgData;
-
-	unsigned int nSize = 0;
+ 
+	size_t nSize = 0;
 	bool bMultipartRelated = false;
 
 	bzero(pMmsMsg, sizeof(MMS_MESSAGE_DATA_S));
